Added Room, Player, and Spawn Managers
So there is a weird thing happening with the player prefab where the transform is like way below them. So anytime you start the game, you like bounce up and then go back down. While you work on the player script, could you please try to fix that? Also I didn't test if GameStateManager stream works, but it's the same as the old project, so yea it'll work. I also created two new states CINEMATICS and DEAD. Cinematics is obv for cutscenes and Dead will be used instead of GameOver when you die in multiplayer. That way, when both players are dead, we can have a script that will set their states to gameover.