Almost done Photonizing Shadowman
Now the only bug is that he has clear preference for the Master Player. Some examples of this preference is if the notmaster is being chased in close pursuit and they run past the master, the monster will go after the master, and the opposite does not happen. Other times, if the notmaster approaches the monster, the monster will actually go for the master instead. This one was much less common, so perhaps with some other debugging I did, that one doesn't happen anymore, but still we should keep our eye out for it